# api/docs/swagger.yaml
definitions:
...
  models.CreateReview:
    properties:
      comment:
        default: very good!
        type: string
      rating:
        default: 4.7
        type: number
    type: object
  models.DeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.ListReviews:
    properties:
      count:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.ReviewModel'
        type: array
    type: object
...
  models.ReviewModel:
    properties:
      comment:
        type: string
      created_at:
        type: string
      establishment_id:
        type: string
      rating:
        type: number
      review_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.StandartError:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
info:
  contact: {}
  description: API for Touristan
  title: Welcome To Booking API
paths:
...
  /v1/review/create:
    post:
      consumes:
      - application/json
      description: Api for creating review
      parameters:
      - description: establishment_id
        in: query
        name: establishment_id
        required: true
        type: string
      - description: user_id
        in: query
        name: user_id
        required: true
        type: string
      - description: Review
        in: body
        name: Review
        required: true
        schema:
          $ref: '#/definitions/models.CreateReview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReviewModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: CREATE REVIEW
      tags:
      - REVIEW
  /v1/review/delete:
    delete:
      consumes:
      - application/json
      description: Api for deleting review by review_id
      parameters:
      - description: review_id
        in: query
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: DELETE REVIEW BY REVIEW_ID
      tags:
      - REVIEW
  /v1/review/list:
    get:
      consumes:
      - application/json
      description: Api for listing reviews by establishment_id
      parameters:
      - description: establishment_id
        in: query
        name: establishment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListReviews'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandartError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandartError'
      security:
      - BearerAuth: []
      summary: LIST REVIEWS BY ESTABLISHMENT_ID
      tags:
      - REVIEW
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"